+++ components/queueManager/Index.ts
@@ -2,10 +2,10 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 export * from './Dashboard/QueueManagerDashboard'
-export * from './QueueManagement'
-export * from './NotManagedCalls'
-export * from './NotManagedCallsFilter'
-export * from './RealTimeManagement'
-export * from './Summary'
+export * from './QueueManagement/QueueManagement'
+export * from './NotManaged/NotManagedCalls'
+export * from './NotManaged/NotManagedCallsFilter'
+export * from './Realtime/RealTimeManagement'
+export * from './Summary/Summary'
 export * from './Monitor'
-export * from './QueueManagementFilterOperators'
+export * from './QueueManagement/QueueManagementFilterOperators'
