+++ components/queueManager/Realtime/RealTimeOperatorsFilter.tsx
@@ -1,63 +1,57 @@
 // Copyright (C) 2023 Nethesis S.r.l.
-// SPDX-License-Identifier: AGPL-3.0-or-later
+// SPDX-License-Identifier: AGPL-3.0-or-lateri
 
 import { ComponentPropsWithRef, forwardRef, useRef } from 'react'
 import classNames from 'classnames'
-import { TextInput } from '../common'
+import { TextInput } from '../../common'
 import { Fragment, useState, useEffect } from 'react'
 import { Dialog, Disclosure, Popover, Transition } from '@headlessui/react'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faChevronDown, faCircleXmark, faXmark } from '@fortawesome/free-solid-svg-icons'
 import { useSelector } from 'react-redux'
-import { RootState } from '../../store'
-import { savePreference } from '../../lib/storage'
-import { DEFAULT_OUTCOME_FILTER, getFilterValues } from '../../lib/queuesLib'
+import { RootState } from '../../../store'
+import { savePreference } from '../../../lib/storage'
+import {
+  DEFAULT_SORT_BY_REALTIME,
+  DEFAULT_STATUS_FILTER_REALTIME,
+  getFilterValuesRealtime,
+} from '../../../lib/queueManager'
 import { useTranslation } from 'react-i18next'
 import { cloneDeep, isEmpty } from 'lodash'
 
-export interface NotManagedCallsFilterProps extends ComponentPropsWithRef<'div'> {
+export interface RealTimeOperatorsFilterProps extends ComponentPropsWithRef<'div'> {
   updateTextFilter: Function
-  updateOutcomeFilter: Function
   updateQueuesFilter: Function
 }
 
-export const NotManagedCallsFilter = forwardRef<HTMLButtonElement, NotManagedCallsFilterProps>(
-  ({ updateTextFilter, updateOutcomeFilter, updateQueuesFilter, className, ...props }, ref) => {
-    const { t } = useTranslation()
+export const RealTimeOperatorsFilter = forwardRef<HTMLButtonElement, RealTimeOperatorsFilterProps>(
+  ({ className, updateTextFilter, updateQueuesFilter, ...props }, ref) => {
     const auth = useSelector((state: RootState) => state.authentication)
-    const [textFilter, setTextFilter] = useState('')
-    const textFilterRef = useRef() as React.MutableRefObject<HTMLInputElement>
-    const [open, setOpen] = useState(false)
-    const [outcome, setOutcome] = useState('')
     const [selectedQueues, setSelectedQueues]: any = useState([])
+    const { t } = useTranslation()
 
     const queueManagerStore = useSelector((state: RootState) => state.queueManagerQueues)
 
-    const outcomeFilter = {
-      id: 'outcome',
-      name: t('Queues.Outcome'),
-      options: [
-        { value: 'lost', label: t('Queues.Not managed') },
-        { value: 'done', label: t('Queues.Managed') },
-        { value: 'all', label: t('Queues.All') },
-      ],
-    }
-
     const queuesFilter = {
       id: 'queues',
-      name: t('Queues.Queues'),
+      name: t('QueueManager.Queues'),
       options: Object.values(queueManagerStore.queues).map((queue: any) => {
         return { value: queue.queue, label: `${queue.name} (${queue.queue})` }
       }),
     }
 
-    function changeTextFilter(event: any) {
-      const newTextFilter = event.target.value
-      setTextFilter(newTextFilter)
+    const [queuesLabel, setQueuesLabel] = useState('')
+    useEffect(() => {
+      if (!isEmpty(selectedQueues)) {
+        setQueuesLabel(selectedQueues.join(', '))
+      } else {
+        // if no queues are selected, it's equivalent to select all of them
+        const allQueueCodes = queuesFilter.options.map((queue: any) => queue.value)
+        setQueuesLabel(allQueueCodes.join(', '))
+      }
+    }, [selectedQueues, queuesFilter.options])
 
-      // notify parent component
-      updateTextFilter(newTextFilter)
-    }
+    const [open, setOpen] = useState(false)
 
     function changeQueuesFilter(event: any) {
       const isChecked = event.target.checked
@@ -71,27 +65,25 @@ export const NotManagedCallsFilter = forwardRef<HTMLButtonElement, NotManagedCal
         newSelectedQueues.splice(index, 1)
         setSelectedQueues(newSelectedQueues)
       }
-      savePreference('queuesSelectedQueues', newSelectedQueues, auth.username)
+      savePreference('realtimeSelectedQueues', newSelectedQueues, auth.username)
 
       // notify parent component
       updateQueuesFilter(newSelectedQueues)
     }
 
-    function changeOutcomeFilter(event: any) {
-      const newOutcome = event.target.id.split('outcome-')[1]
-      setOutcome(newOutcome)
-      savePreference('queuesOutcomeFilter', newOutcome, auth.username)
+    const [textFilter, setTextFilter] = useState('')
+    const textFilterRef = useRef() as React.MutableRefObject<HTMLInputElement>
+    function changeTextFilter(event: any) {
+      const newTextFilter = event.target.value
+      setTextFilter(newTextFilter)
 
-      // notify parent component
-      updateOutcomeFilter(newOutcome)
+      // update operators (notify parent component)
+      updateTextFilter(newTextFilter)
     }
 
     // retrieve filter values from local storage
-
     useEffect(() => {
-      const filterValues = getFilterValues(auth.username)
-      setOutcome(filterValues.outcome)
-      updateOutcomeFilter(filterValues.outcome)
+      const filterValues = getFilterValuesRealtime(auth.username)
 
       if (isEmpty(filterValues.selectedQueues)) {
         // select all queues
@@ -107,37 +99,13 @@ export const NotManagedCallsFilter = forwardRef<HTMLButtonElement, NotManagedCal
       }
     }, [])
 
-    // outcome label
-
-    const [outcomeLabel, setOutcomeLabel] = useState('')
-    useEffect(() => {
-      const found = outcomeFilter.options.find((option) => option.value === outcome)
-
-      if (found) {
-        setOutcomeLabel(found.label)
-      }
-    }, [outcome, outcomeFilter.options])
-
-    // queues label
-
-    const [queuesLabel, setQueuesLabel] = useState('')
-    useEffect(() => {
-      if (!isEmpty(selectedQueues)) {
-        setQueuesLabel(selectedQueues.join(', '))
-      } else {
-        // if no queues are selected, it's equivalent to select all of them
-        const allQueueCodes = queuesFilter.options.map((queue: any) => queue.value)
-        setQueuesLabel(allQueueCodes.join(', '))
-      }
-    }, [selectedQueues, queuesFilter.options])
-
     const resetFilters = () => {
       setTextFilter('')
-      updateTextFilter('') // notify parent component
+      savePreference('realtimeStatusFilter', DEFAULT_STATUS_FILTER_REALTIME, auth.username)
+      savePreference('realtimeStatusSortBy', DEFAULT_SORT_BY_REALTIME, auth.username)
 
-      setOutcome(DEFAULT_OUTCOME_FILTER)
-      updateOutcomeFilter(DEFAULT_OUTCOME_FILTER) // notify parent component
-      savePreference('queuesOutcomeFilter', DEFAULT_OUTCOME_FILTER, auth.username)
+      // notify parent component
+      updateTextFilter('')
 
       // select all queues
       const allQueueCodes = Object.values(queueManagerStore.queues).map((queue: any) => {
@@ -145,7 +113,7 @@ export const NotManagedCallsFilter = forwardRef<HTMLButtonElement, NotManagedCal
       })
       setSelectedQueues(allQueueCodes)
       updateQueuesFilter(allQueueCodes)
-      savePreference('queuesSelectedQueues', allQueueCodes, auth.username)
+      savePreference('realtimeSelectedQueues', allQueueCodes, auth.username)
     }
 
     const clearTextFilter = () => {
@@ -199,73 +167,6 @@ export const NotManagedCallsFilter = forwardRef<HTMLButtonElement, NotManagedCal
 
                     {/* Filters (mobile) */}
                     <form className='mt-4'>
-                      {/* outcome filter (mobile) */}
-                      <Disclosure
-                        as='div'
-                        key={outcomeFilter.name}
-                        className='border-t px-4 py-6 border-gray-200 dark:border-gray-700'
-                      >
-                        {({ open }) => (
-                          <>
-                            <h3 className='-mx-2 -my-3 flow-root'>
-                              <Disclosure.Button className='flex w-full items-center justify-between px-2 py-3 text-sm bg-white text-gray-400 dark:bg-gray-900 dark:text-gray-500'>
-                                <span className='font-medium text-gray-900 dark:text-gray-100'>
-                                  {outcomeFilter.name}
-                                </span>
-                                <span className='ml-6 flex items-center'>
-                                  <FontAwesomeIcon
-                                    icon={faChevronDown}
-                                    className={classNames(
-                                      open ? '-rotate-180' : 'rotate-0',
-                                      'h-3 w-3 transform',
-                                    )}
-                                    aria-hidden='true'
-                                  />
-                                </span>
-                              </Disclosure.Button>
-                            </h3>
-                            <Disclosure.Panel className='pt-6'>
-                              <fieldset>
-                                <legend className='sr-only'>{outcomeFilter.name}</legend>
-                                <div className='space-y-4'>
-                                  {outcomeFilter.options.map((option) => (
-                                    <div key={option.value}>
-                                      {option.value.startsWith('divider') ? (
-                                        <div className='relative'>
-                                          <div
-                                            className='absolute inset-0 flex items-center'
-                                            aria-hidden='true'
-                                          >
-                                            <div className='w-full border-t border-gray-300 dark:border-gray-600' />
-                                          </div>
-                                          <div className='relative flex justify-center'></div>
-                                        </div>
-                                      ) : (
-                                        <div className='flex items-center'>
-                                          <input
-                                            id={`outcome-${option.value}`}
-                                            name={`filter-${outcomeFilter.id}`}
-                                            type='radio'
-                                            defaultChecked={option.value === outcome}
-                                            onChange={changeOutcomeFilter}
-                                            className='h-4 w-4 border-gray-300 text-primary focus:ring-primaryLight dark:border-gray-600 dark:text-primary dark:focus:ring-primaryDark'
-                                          />
-                                          <label
-                                            htmlFor={`outcome-${option.value}`}
-                                            className='ml-3 block text-sm font-medium text-gray-700 dark:text-gray-200'
-                                          >
-                                            {option.label}
-                                          </label>
-                                        </div>
-                                      )}
-                                    </div>
-                                  ))}
-                                </div>
-                              </fieldset>
-                            </Disclosure.Panel>
-                          </>
-                        )}
-                      </Disclosure>
                       {/* queues filter (mobile) */}
                       <Disclosure
                         as='div'
@@ -330,13 +231,13 @@ export const NotManagedCallsFilter = forwardRef<HTMLButtonElement, NotManagedCal
           <div className='mx-auto text-center'>
             <section aria-labelledby='filter-heading' className='pb-6'>
               <h2 id='filter-heading' className='sr-only'>
-                {t('Queues.Calls filters')}
+                {t('QueueManager.Realtime operators filters')}
               </h2>
 
-              <div className='flex items-center'>
+              <div className='flex items-center space-x-8'>
                 <div className='flex items-center'>
                   <TextInput
-                    placeholder='Filter calls'
+                    placeholder='Filter operators'
                     className='max-w-sm'
                     value={textFilter}
                     onChange={changeTextFilter}
@@ -347,76 +248,9 @@ export const NotManagedCallsFilter = forwardRef<HTMLButtonElement, NotManagedCal
                   />
                 </div>
 
-                <div className='flex ml-8'>
+                <div className='flex ml-4'>
                   <Popover.Group className='hidden sm:flex sm:items-baseline sm:space-x-4'>
-                    {/* outcome filter */}
-                    <Popover
-                      as='div'
-                      key={outcomeFilter.name}
-                      id={`desktop-menu-${outcomeFilter.id}`}
-                      className='relative inline-block text-left shrink-0'
-                    >
-                      <div>
-                        <Popover.Button className='px-3 py-2 text-sm leading-4 p-2 rounded border shadow-sm border-gray-300 bg-white text-gray-700 hover:bg-gray-100 focus:ring-primaryLight dark:border-gray-600 dark:bg-gray-900 dark:text-gray-100 dark:hover:bg-gray-800 dark:focus:ring-primaryDark group inline-flex items-center justify-center font-medium  hover:text-gray-900 dark:hover:text-gray-100'>
-                          <span>{outcomeFilter.name}</span>
-                          <FontAwesomeIcon
-                            icon={faChevronDown}
-                            className='ml-2 h-3 w-3 flex-shrink-0 text-gray-400 group-hover:text-gray-500 dark:text-gray-500 dark:group-hover:text-gray-400'
-                            aria-hidden='true'
-                          />
-                        </Popover.Button>
-                      </div>
-
-                      <Transition
-                        as={Fragment}
-                        enter='transition ease-out duration-100'
-                        enterFrom='transform opacity-0 scale-95'
-                        enterTo='transform opacity-100 scale-100'
-                        leave='transition ease-in duration-75'
-                        leaveFrom='transform opacity-100 scale-100'
-                        leaveTo='transform opacity-0 scale-95'
-                      >
-                        <Popover.Panel className='absolute right-0 z-10 mt-2 origin-top-right rounded-md min-w-max p-4 shadow-2xl ring-1 focus:outline-none ring-opacity-5 bg-white ring-black dark:ring-opacity-5 dark:bg-gray-900 dark:ring-gray-700'>
-                          <form className='space-y-4'>
-                            {outcomeFilter.options.map((option) => (
-                              <div key={option.value}>
-                                {option.value.startsWith('divider') ? (
-                                  <div className='relative'>
-                                    <div
-                                      className='absolute inset-0 flex items-center'
-                                      aria-hidden='true'
-                                    >
-                                      <div className='w-full border-t border-gray-300 dark:border-gray-600' />
-                                    </div>
-                                    <div className='relative flex justify-center'></div>
-                                  </div>
-                                ) : (
-                                  <div className='flex items-center'>
-                                    <input
-                                      id={`outcome-${option.value}`}
-                                      name={`filter-${outcomeFilter.id}`}
-                                      type='radio'
-                                      defaultChecked={option.value === outcome}
-                                      onChange={changeOutcomeFilter}
-                                      className='h-4 w-4 border-gray-300 text-primary focus:ring-primaryLight dark:border-gray-600 dark:text-primary dark:focus:ring-primaryDark'
-                                    />
-                                    <label
-                                      htmlFor={`outcome-${option.value}`}
-                                      className='ml-3 block text-sm font-medium text-gray-700 dark:text-gray-200'
-                                    >
-                                      {option.label}
-                                    </label>
-                                  </div>
-                                )}
-                              </div>
-                            ))}
-                          </form>
-                        </Popover.Panel>
-                      </Transition>
-                    </Popover>
-
                     {/* queues filter */}
-
                     <Popover
                       as='div'
                       key={queuesFilter.name}
@@ -488,19 +322,6 @@ export const NotManagedCallsFilter = forwardRef<HTMLButtonElement, NotManagedCal
                   </h3>
                   {/* separator */}
                   <div aria-hidden='true' className='h-5 w-px block bg-gray-300 dark:bg-gray-600' />
-                  {/* outcome */}
-                  <div className='mt-0'>
-                    <div className='-m-1 flex flex-wrap items-center'>
-                      <span className='m-1 inline-flex items-center rounded-full border py-1.5 px-3 text-sm font-medium border-gray-300 text-gray-700 dark:border-gray-600 dark:text-gray-100'>
-                        <span>
-                          <span className='text-gray-600 dark:text-gray-300'>
-                            {t('Queues.Outcome')}:
-                          </span>{' '}
-                          {outcomeLabel}
-                        </span>
-                      </span>
-                    </div>
-                  </div>
                   {/* queues */}
                   <div className='mt-0'>
                     <div className='-m-1 flex flex-wrap items-center'>
@@ -540,4 +361,4 @@ export const NotManagedCallsFilter = forwardRef<HTMLButtonElement, NotManagedCal
   },
 )
 
-NotManagedCallsFilter.displayName = 'NotManagedCallsFilter'
+RealTimeOperatorsFilter.displayName = 'RealTimeOperatorsFilter'
