+++ pages/queueManager.tsx
@@ -6,10 +6,10 @@ import { useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import classNames from 'classnames'
 import { QueueManagerDashboard } from '../components/queueManager/Dashboard/QueueManagerDashboard'
-import { QueueManagement } from '../components/queueManager/QueueManagement'
-import { NotManagedCalls } from '../components/queueManager/NotManagedCalls'
-import { RealTimeManagement } from '../components/queueManager/RealTimeManagement'
-import { Summary } from '../components/queueManager/Summary'
+import { QueueManagement } from '../components/queueManager/QueueManagement/QueueManagement'
+import { NotManagedCalls } from '../components/queueManager/NotManaged/NotManagedCalls'
+import { RealTimeManagement } from '../components/queueManager/Realtime/RealTimeManagement'
+import { Summary } from '../components/queueManager/Summary/Summary'
 import { Monitor } from '../components/queueManager/Monitor'
 import { useSelector } from 'react-redux'
 import { RootState, store } from '../store'
